Moving isOptionalString helper-function to TypeNameAnalyser
